requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.3
celery==5.5.3
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
Django==5.2.8
django-celery-beat==2.8.1
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.1
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.52
python-crontab==3.3.0
python-dateutil==2.9.0.post0
redis==7.0.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
